Upgrade to Java 25.

+++ advanced/gis/giscorefx/src/main/java/org/arakhne/afc/gis/ui/drawers/AbstractGISEditableDrawer.java
@@ -5,7 +5,7 @@
  * Copyright (c) 2000-2012 Stephane GALLAND.
  * Copyright (c) 2005-10, Multiagent Team, Laboratoire Systemes et Transports,
  *                        Universite de Technologie de Belfort-Montbeliard.
- * Copyright (c) 2013-2023 The original authors and other contributors.
+ * Copyright (c) 2013-2026 The original authors and other contributors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,12 +20,11 @@
 
 package org.arakhne.afc.gis.ui.drawers;
 
-import org.eclipse.xtext.xbase.lib.Pure;
-
 import org.arakhne.afc.gis.primitive.FlagContainer;
 import org.arakhne.afc.gis.primitive.GISEditable;
 import org.arakhne.afc.nodefx.Drawer;
 import org.arakhne.afc.vmutil.ColorNames;
+import org.eclipse.xtext.xbase.lib.Pure;
 
 /** Abstract drawer of a GIS editable.
  *
@@ -38,7 +37,7 @@ import org.arakhne.afc.vmutil.ColorNames;
  */
 public abstract class AbstractGISEditableDrawer<T extends GISEditable & FlagContainer>  implements Drawer<T> {
 
-	private static final int SELECTION_COLOR = ColorNames.getColorFromName("lightcoral"); //$NON-NLS-1$
+	private static final int SELECTION_COLOR = ColorNames.getColorFromName("lightcoral").intValue(); //$NON-NLS-1$
 
 	/** Replies the given element is selected.
 	 *
