+++ core/maths/mathgeom/src/main/java/org/arakhne/afc/math/geometry/AbstractGeomFactoryBase.java
@@ -5,7 +5,7 @@
  * Copyright (c) 2000-2012 Stephane GALLAND.
  * Copyright (c) 2005-10, Multiagent Team, Laboratoire Systemes et Transports,
  *                        Universite de Technologie de Belfort-Montbeliard.
- * Copyright (c) 2013-2023 The original authors and other contributors.
+ * Copyright (c) 2013-2026 The original authors and other contributors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -74,7 +74,7 @@ public abstract class AbstractGeomFactoryBase implements GeomFactoryBase {
 		if (distance == null || Double.isNaN(distance.doubleValue())) {
 			globalSplineApproximation = GeomConstants.SPLINE_APPROXIMATION_RATIO;
 		} else {
-			globalSplineApproximation = Math.max(0, distance);
+			globalSplineApproximation = Math.max(0, distance.doubleValue());
 		}
 	}
 
@@ -90,7 +90,7 @@ public abstract class AbstractGeomFactoryBase implements GeomFactoryBase {
 
 	@Override
 	public void setSplineApproximationRatio(Double distance) {
-		if (distance == null || Double.isNaN(distance)) {
+		if (distance == null || Double.isNaN(distance.doubleValue())) {
 			this.splineApproximation = null;
 		} else {
 			this.splineApproximation = distance;
